Tying screens to bases
git-svn-id: https://openxcom.svn.sourceforge.net/svnroot/openxcom/trunk@50 11c4a5ed-7179-4546-81a6-367b3be9d812

+++ SurfaceSet.cpp
@@ -133,7 +133,7 @@ void SurfaceSet::loadDat(string filename)
 	streamoff size = imgFile.tellg();
 	imgFile.seekg(0, ios::beg);
 
-	_nframe = size / (_width * _height);
+	_nframe = (int)size / (_width * _height);
 
 	for (int i = 0; i < _nframe; i++)
 	{
